+++ Assets/Scripts/Combat/Enemy.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public abstract class Enemy : MonoBehaviour, IDamageable {
-  public float health;
+  public int health;
   
   public abstract void ActAggressive();
   public abstract void ActDefensive();
-  public abstract void TakeDamage(float damage);
+  public abstract void TakeDamage(int damage);
 }
